fix(tuev): update TUEV event training script for class count tensor type
- Changed the tensor type for class counts in the TUEV event training script to `torch.long` to ensure compatibility and prevent potential issues during training.
- This update aims to enhance the robustness of the training process by ensuring that class counts are correctly calculated and represented, contributing to improved model performance.

+++ experiments/eegpt_linear_probe/train_tuev_events.py
@@ -415,7 +415,7 @@ def main(args):
     with open(train_dataset.cache_dir / train_dataset.split / "index.json", "r") as f:
         index_data = json.load(f)
     train_labels = [seg["label"] for seg in index_data["segments"]]
-    class_counts = torch.bincount(torch.tensor(train_labels), minlength=6)
+    class_counts = torch.bincount(torch.tensor(train_labels, dtype=torch.long), minlength=6)
     print(f"Class distribution (natural): {class_counts.tolist()}")
     print("Using natural distribution - NO rebalancing")
